web/double-check/solver/requirements.txt
requests == 2.27.1
pyjwt == 0.4.3